pyproject.toml
[tool.poetry]
name = "incident-assistant"
version = "0.1.0"
description = "Цифровой помощник для управления инцидентами на НПЗ"
authors = ["Your Name <your@email.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.44"}
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"
pydantic-settings = "^2.11.0"
asyncpg = "^0.30.0"
email-validator = "^2.3.0"
bcrypt = "^5.0.0"
passlib = "^1.7.4"
argon2-cffi = "^25.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"